SystemUI: Do not allow to access quick settings on secure lockscreen
We have a lot security guys as user due of using the phones as business phones...that
said a lot complains go into the direction of google how it is possible to have
access to quick settings on a secure lockscreen which is absolutely understandable
due that you are able to control Wifi BT AirplaneMode data etc.

This patch takes care on this problem and controls this behaviour.

ATTENTION: Due that this is really an UX derp for an average user this feature is
enabled by default now.

/-/

Changes apported to original patch and reflected in the edited commit message above:
* This has been integrated and is hardcoded, cause this is the behaviour it should have.
  Thanks Larss for this.

Signed-off-by: @BigBrother1984 <carlosavignano@aospa.co>
Change-Id: Ib437802ae8687dc8e99122a40e285bdd2df3e04c

Conflicts:
	packages/SystemUI/src/com/android/systemui/statusbar/phone/NotificationPanelView.java