Frameworks: Actually start to use HEADS_UP_REQUESTED extra
Prepairing slim heads up feature, we do this here to allow apps to
send a heads up notification no matter what.

Google introduced this flag but is not using it in the actual API

Take HEADS_UP_REQUESTED into account to indicate that this notification
is a good heads up candidate.

How to use it you can see in email or mms app

Change-Id: I4a29404122ae12ab9b95e1d223911078a360b44e

Conflicts:
	packages/SystemUI/src/com/android/systemui/statusbar/BaseStatusBar.java