Fix screen off finished notification to notify after the screen is turned off.
This patch can prevent flickering of UI components before the screen off
animation is played. Reproducing steps and effects are explained at
https://code.google.com/p/android/issues/detail?id=143650.

To prevent flickering, Notifier.onInteractiveStateChangeFinished(false)
should be called after the screen is turned off.
In the KitKat, Notifier.onGoToSleepFinished() is called correctly
after the screen off animation is finished.

However, in the Lollipop, during the screen off process,
setInteractiveStateLocked(false, reason) is called twice
from the goToSleepNoUpdateLocaked and the reallyGoToSleepNoUpdateLocked.
But, the second call is ignored, because the mInteractive and
mInteractiveChainging is changed after the first call.
The finishInteractiveStateChangedLocked also processed after the first call,
that is the root cause of flickering effect.

This patch removes the first call of setInteractiveStateLocked,
and modify the second call to using proper reason.
This makes sure that the onInteractiveStateChangeFinished
is called after the screen is turned off.

Change-Id: I4787edf0c2feb2155994085d7c89149970b4f57c