SystemUI: improve DISMISS_KEYGUARD_SECURELY_ACTION behavior
When the screen is not yet turned on, calls to dismiss() will never
actually dismiss the keyguard since it goes through the bouncer, which
checks if the screen is on before executing its logic. If we receive the
intent to dismiss the keyguard while the screen is off, wait until its
turned on and then execute the dismiss.

Change-Id: I652458c639230bc7fe32a955e103e40c27289b4d
Signed-off-by: Roman Birg <roman@cyngn.com>

Conflicts:
	packages/SystemUI/src/com/android/systemui/keyguard/KeyguardViewMediator.java